change: move room to new folder
